Use some shared utilities
